ensureConnection - fixed rejection handling (#2316)
Hi, I was trying to create a scenario that we can not establish
BillingClient connection. To do that, I simply changed `initConnection`
as below:

```
fun initConnection(promise: Promise) {
         Log.i(TAG, "Google Play Services are not available on this device")
         promise.safeReject(PromiseUtils.E_NOT_PREPARED, "Google Play Services are not available on this device")
         return
}
```
   
After observing that scenario, I noticed that `ensureConnection` is not
parsing the rejection parameters properly and it's only logging this
info in the logcat, which is quite dangerous because it's happening
silently and probably none of us is aware of this issue so far.

When it's rejected, there is only 1 item in `it` array and here is the
output of `it[0]`:
![Screen Shot 2023-03-16 at 13 03
08](https://user-images.githubusercontent.com/16003411/225621284-bec5c01c-7ae1-4cd0-bc6a-8a9d2c8d0b00.png)

I tried to fix the parameter handling and rejected the case when the
parameters are incorrect instead of just logging it. So that, we will be
aware of if anything goes wrong.

---------

Co-authored-by: Andres Aguilar <andresesfm@gmail.com>